feat: GithubLogoImage 분리
